Feat: UserChoiceResponse, TimeRanking dto 변경
Former-commit-id: de49ca26f35456b2b66aaf93c6019f9df7bba31e

+++ android/app/src/main/java/com/ssafy/firstproject/data/model/response/TimeRanking.kt
@@ -1,6 +1,6 @@
 package com.ssafy.firstproject.data.model.response
 
 data class TimeRanking(
-    val dwellTime: Int,
+    val dwellTime: Long,
     val nickname: String
 )
